Add types to protected-route
